refactor: useTermsCheck 훅 분리 (#81)
